finalize seeding method
